AOI change and function add
removed redundant code about converting the aoi to geojson  which was leftover when mistakenly trying to use sentinelsat. modification of the rio.open band code to make this into a function since it needs to be used more than once on different bands. Leaves open to make easier to add other indices at a later date.