ndvi calculation mods
change to NIR and RED variable outputs to ensure they are just the array without the affine transformation, since this created an error when trying to calculate ndvi which wouldn't do band maths on a tuple.
Also a change to the ndvi formula to get rid of problems of dividing zero. Still an issue that the ndvi output needs to be georeferenced with the same info as the orignal raster.
Also a couple of typo modifications.