ndvi.png output code added
Code added to output a png of the NDVI map with title and labels. Also minor changes, setting a new variable with a crs for use in several lines of code for consistency. minor changes to some text and to ensure files are saved with unique file names for searching for different squares